Added language loading on language change.
There is a memory issue, but it will force a refresh if the error occurs. This fixes the issue, but is obviously an unoptimal solution.

There is code copy pasted from preloadStore into inventory. A better solution would be to have a single function both could use, but I was unable to get one working.